电流限幅 60A + DRV8353 停机保护 (PHU 风格三阶段)
## 电流限幅放开到 60A
- INC_PID_SPEED_LIMIT: 10A → 60A (额定 65Nm 需 ~22A)
- DEFAULT_MAX_CURRENT: 25A → 60A
- OverCurrent 保护: 58A → 64A (uint16_t max)
- LOCKED_MOTOR_CURRENT: 15A → 75A (堵转阈值跟随放宽)
- set_torque_ref / CurrentLoopSmoothRun: int16_t → int32_t (防溢出)
- Clarke/Park 变换: int32_t → int64_t 中间变量 (防 60A 溢出)

## DRV8353 停机保护 (对齐 PHU 三阶段方案)
**核心思路**: 不立即短路刹车, 先零电压矢量让电机自由滑行,
速度降到安全阈值后再短路, 最后关 MOE 进入高阻。

阶段1: 零电压滑行 (最长 500ms)
  - 保持 foc_run, 电流环继续运行
  - 清零 I_q_ref/I_d_ref/velocity_ref
  - SVPWM 输出零电压矢量 (CCR ≈ PWM_T/4)
  - 退出条件: |dtheta_mech| < 2rpm 载端 (51200) 或超时

阶段2: 低速短路刹车 (固定 30ms)
  - foc_run=0, CCR=0 三相下管全开
  - 速度已 <2rpm, 反电动势仅 ~0.4V, 刹车电流 <6A 安全

阶段3: 高阻关断
  - TIM1->BDTR &= ~MOE, CCER &= ~0x1555
  - 清理 velocity_ref/I_q_ref/position_ref 等运行数据

## 配套改动
- enable0 命令: 移除旧的高速保护 + 30ms 预衰减逻辑,
  直接调 fault_safe_shutdown() 走统一状态机
- CheckAndHandleAllFaultBits: 斜坡减速完成后不再立即关 foc_run,
  改为调 fault_safe_shutdown() 进入滑行阶段

## 损坏机理 vs PHU 安全机理
旧方案 (直接 CCR=0): 高速反电动势 ~10V → 短路电流 130A → DRV 击穿
新方案 (滑行后刹车): 滑行段 CCR=PWM_T/4 上下管正常交替, 无短路;
                    刹车段速度已低, I_brake ≈ V_bemf/Rs < 6A

## 其他保护层 (保留)
- 严重故障 (OVP/OC/OT): 50ms 快速斜坡 → 进入停机状态机
- 一般故障: 150ms 保守斜坡 → 进入停机状态机
- 死区加倍: 50ns → 100ns (降低 shoot-through 风险)
- BKIN 硬件保护: nFAULT 拉低自动关 MOE

+++ foc/foc_fast/foc_api.c
@@ -79,11 +79,11 @@ int32_t get_velocity_lim(void) {
     return 0;
 }
 
-int16_t set_torque_ref(int16_t TorRef) {
+int32_t set_torque_ref(int32_t TorRef) {
     if (controller_eyou.ServoState.Bit.PdoRefreshing &&
         (controller_eyou.controller_mode == PROFILE_TORQUE_MODE ||
          controller_eyou.controller_mode == CYCLIC_SYNC_TORQUE_MODE)) {
-        int16_t max = (int16_t)controller_eyou.FlashData.MaxCurrent;
+        int32_t max = (int32_t)controller_eyou.FlashData.MaxCurrent;
         if (TorRef >  max) TorRef =  max;
         if (TorRef < -max) TorRef = -max;
         controller_eyou.I_q_ref = TorRef;
@@ -91,8 +91,8 @@ int16_t set_torque_ref(int16_t TorRef) {
     return 1;
 }
 
-int16_t set_torque_ref_loop(int16_t TorRef) {
-    int16_t max = (int16_t)controller_eyou.FlashData.MaxCurrent;
+int32_t set_torque_ref_loop(int32_t TorRef) {
+    int32_t max = (int32_t)controller_eyou.FlashData.MaxCurrent;
     if (TorRef >  max) TorRef =  max;
     if (TorRef < -max) TorRef = -max;
     controller_eyou.I_q_ref = TorRef;
